+++ src/test/Fake.Core.UnitTests/Fake.Core.Target.fs
@@ -52,7 +52,7 @@ module TestTarget =
 let tests =
   testList "Fake.Core.Target.Tests" (
     [
-    Fake.ContextHelper.fakeContextTestCaseAssertTime (TimeSpan.FromSeconds 10.0) "basic performance #2036" <| fun _ ->
+    Fake.ContextHelper.fakeContextTestCaseAssertTime (TimeSpan.FromSeconds 13.0) "basic performance #2036" <| fun _ ->
         let counter = 2500
         CoreTracing.setTraceListeners [] // silence
         let all_Pipelines = Dictionary<string,TestTarget>(System.StringComparer.OrdinalIgnoreCase :> IEqualityComparer<string>)